Seperated list function into lister local module.
